pyproject.toml
[project]
name = "devlog"
version = "0.1.0"
description = "My Own Blog with Dev story"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.8.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "google-auth>=2.40.3",
    "ipython>=9.4.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-stubs>=5.2.2",
    "drf-spectacular>=0.28.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
]

[tool.mypy]
python_version = "3.13"                # 코드가 Python 3.13 기준으로 작성되었음을 명시
plugins = ["mypy_django_plugin.main"]  # Django 프로젝트용 mypy plugin 적용
ignore_missing_imports = true          # 설치되지 않은 패키지 import 에러 무시
disable_error_code = ["var-annotated"] # 특정 오류 코드(var-annotated) 무시
explicit_package_bases = true          # 명시적 패키지 베이스 사용으로 모듈 충돌 해결

[tool.django-stubs]
django_settings_module = 'config.settings.local'

[tool.isort]
# black과 포맷 스타일을 맞춤
profile="black"

